feat: Change data type

+++ mandatory/ft_cube3d.h
@@ -46,14 +46,14 @@ typedef struct s_rcast
 	int		x;
 	double	ray_dir_x;
 	double	ray_dir_y;
-	double	map_x;
-	double	map_y;
+	int		map_x;
+	int		map_y;
 	double	delta_dist_x;
 	double	delta_dist_y;
 	double	side_dist_x;
 	double	side_dist_y;
-	double	step_x;
-	double	step_y;
+	int		step_x;
+	int		step_y;
 	int		side;
 	double	perp_wall_dist;
 	int		line_height;
